rename hotkey-compare > normalize-hotkey
