+++ src/normalize-hotkey.ts
