+++ test/test-normalize-hotkey.js
